First version of the new temp temp file handling

+++ src/graphconstructor/constructor.cpp
@@ -104,11 +104,11 @@ int main(int argc, char * argv[])
 			"integer",
 			cmd);
 
-		TCLAP::ValueArg<unsigned int> abundance("a",
+		TCLAP::ValueArg<size_t> abundance("a",
 			"abundance",
 			"Vertex abundance threshold",
 			false,
-			UINT32_MAX,
+			UINT64_MAX,
 			"integer",
 			cmd);
